requirements.txt
fastapi
uvicorn
pydantic
matplotlib
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
